convert struct impl to use structured format
